api/requirements.txt
flask
flask_socketio
eventlet
paho-mqtt
requests
